refactor: Move showPopover() into its own file
